Refactor LTI iframe checker
Seems that all of the random `Cannot read property 'document' of null`
errors were caused by how we are trying to access the needed iframes. If
any issue happened while checking the iframes, any future cypress
retry would give the non-useful error message.
With this other way it seems that retrials are possible. Hopefully it
means that we will start seeing valid failure messages if anything goes
bad with LTI.